requirements.txt
faiss==1.5.3
fonttools==4.55.2
hdbscan==0.8.40
joblib==1.4.2
keras==3.7.0
matplotlib==3.10.1
numpy==2.2.4
opencv_python==4.10.0.84
openpyxl==3.1.5
pandas==2.2.3
Pillow==11.1.0
rasterio==1.4.3
scikit_learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
segment_anything==1.0
spectral==0.23.1
tensorflow==2.18.0
torch==2.5.1
umap==0.1.1
